requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
numpy>=1.26.0
SQLAlchemy>=2.0.25
ofxparse>=0.21
python-dateutil>=2.9.0
altair>=5.3.0
pydantic>=2.8.2
